pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tag-dqn"
version = "0.1.0"
description = "Term Analysis with Graph Deep Q-Networks"
readme = "README.md"
authors = [
    { name = "Milan Ding", email = "milan.ding15@imperial.ac.uk" }
]
requires-python = ">=3.11.11"
classifiers = [
    "Programming Language :: Python :: 3.11"
]

dependencies = [
    "numpy >=2.0.1,<3",
    "scipy >=1.15.1,<2",
    "matplotlib >=3.10.0,<4",
    "pandas >=2.2.3,<3",
    "openpyxl >=3.1.5,<4",
    "torch >=2.4.1,<3",
    "torch_geometric >=2.6.1,<3",
    "statsmodels >=0.14.4,<1",
    "PyYAML >=6.0.2,<7",
    "tqdm >=4.67.1,<5",
    "cachetools >= 5.5.2,<6"
]

[tool.setuptools.package-data]
tag_dqn = ["pkg_data/reward.pth"]

[tool.setuptools.packages.find]
include = ["tag_dqn*"]
